Many small but important improvements
- fixed negative row index bug in tableCellClick
- improved error handling (more user friendly now)
- refactored watcher close by code
- updated log file extension to .log
- added build command to Makefile
- updated README
- updated dependency packages

+++ modules/connections/models/fetch_item.go
@@ -25,9 +25,9 @@ type FetchItem struct {
 				Realm             string `json:"realm"`
 			} `json:"account"`
 			Price struct {
-				Type     string `json:"type"`
-				Amount   int    `json:"amount"`
-				Currency string `json:"currency"`
+				Type     string  `json:"type"`
+				Amount   float64 `json:"amount"`
+				Currency string  `json:"currency"`
 			} `json:"price"`
 		} `json:"listing"`
 		Item struct {
